Update README.md
🚀 Exciting News! Just Launched My Second Web Development Project! 🌐

Hey GitHub community! 👋 I'm thrilled to share my latest achievement—a web development project created with HTML! 🖥️✨ Dive into the details of my Admission Form and join me on this coding journey. 🎓📝

🔍 Explore the Admission Form Project: [Admission Form]

🌱 About the Project:
In this project, I've harnessed the power of HTML tags to craft a user-friendly Admission Form. As a beginner in web development, I'm excited to showcase my progress and share a practical application of HTML in creating a meaningful web page. 💻💡